+++ service/src/main/java/kitchenpos/service/menugroup/dto/MenuGroupRequest.java
@@ -1,0 +1,20 @@
+package kitchenpos.service.menugroup.dto;
+
+import kitchenpos.menugroup.domain.MenuGroup;
+
+public class MenuGroupRequest {
+    private String name;
+
+    public MenuGroupRequest() {}
+    public MenuGroupRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public MenuGroup toMenuGroup() {
+        return new MenuGroup(this.name);
+    }
+}
