+++ service/src/main/java/kitchenpos/service/menugroup/dto/MenuGroupResponse.java
@@ -1,0 +1,21 @@
+package kitchenpos.service.menugroup.dto;
+
+import kitchenpos.menugroup.domain.MenuGroup;
+
+public class MenuGroupResponse {
+    private Long id;
+    private String name;
+
+    public MenuGroupResponse(MenuGroup menuGroup) {
+        this.id = menuGroup.getId();
+        this.name = menuGroup.getName();
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
