+++ service/src/main/java/kitchenpos/service/table/dto/OrderTableResponse.java
@@ -1,0 +1,27 @@
+package kitchenpos.service.table.dto;
+
+import kitchenpos.table.domain.OrderTable;
+
+public class OrderTableResponse {
+    private Long id;
+    private int numberOfGuests;
+    private boolean isEmpty;
+
+    public OrderTableResponse(OrderTable orderTable) {
+        this.id = orderTable.getId();
+        this.numberOfGuests = orderTable.getNumberOfGuests();
+        this.isEmpty = orderTable.isEmpty();
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public int getNumberOfGuests() {
+        return numberOfGuests;
+    }
+
+    public boolean isEmpty() {
+        return isEmpty;
+    }
+}
