+++ service/src/main/java/kitchenpos/service/table/dto/OrderTableUpdateEmptyRequest.java
@@ -1,0 +1,16 @@
+package kitchenpos.service.table.dto;
+
+public class OrderTableUpdateEmptyRequest {
+    private boolean empty;
+
+    public OrderTableUpdateEmptyRequest() {
+    }
+
+    public OrderTableUpdateEmptyRequest(boolean empty) {
+        this.empty = empty;
+    }
+
+    public boolean isEmpty() {
+        return empty;
+    }
+}
