--- src/main/java/kitchenpos/menugroup/dto/MenuGroupResponse.java
@@ -1,21 +1,0 @@
-package kitchenpos.menugroup.dto;
-
-import kitchenpos.menugroup.domain.MenuGroup;
-
-public class MenuGroupResponse {
-    private Long id;
-    private String name;
-
-    public MenuGroupResponse(MenuGroup menuGroup) {
-        this.id = menuGroup.getId();
-        this.name = menuGroup.getName();
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
