--- src/main/java/kitchenpos/order/dto/OrderLineItemRequest.java
@@ -1,22 +1,0 @@
-package kitchenpos.order.dto;
-
-public class OrderLineItemRequest {
-    private long menuId;
-    private long quantity;
-
-    public OrderLineItemRequest() {
-    }
-
-    public OrderLineItemRequest(long menuId, long quantity) {
-        this.menuId = menuId;
-        this.quantity = quantity;
-    }
-
-    public long getMenuId() {
-        return menuId;
-    }
-
-    public long getQuantity() {
-        return quantity;
-    }
-}
