--- src/main/java/kitchenpos/table/dto/OrderTableUpdateEmptyRequest.java
@@ -1,16 +1,0 @@
-package kitchenpos.table.dto;
-
-public class OrderTableUpdateEmptyRequest {
-    private boolean empty;
-
-    public OrderTableUpdateEmptyRequest() {
-    }
-
-    public OrderTableUpdateEmptyRequest(boolean empty) {
-        this.empty = empty;
-    }
-
-    public boolean isEmpty() {
-        return empty;
-    }
-}
