--- src/main/java/kitchenpos/table/dto/OrderTableUpdateNumberOfGuestsRequest.java
@@ -1,20 +1,0 @@
-package kitchenpos.table.dto;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-public class OrderTableUpdateNumberOfGuestsRequest {
-    private int numberOfGuests;
-
-    public OrderTableUpdateNumberOfGuestsRequest() {
-    }
-
-    @JsonCreator
-    public OrderTableUpdateNumberOfGuestsRequest(@JsonProperty("numberOfGuests") int numberOfGuests) {
-        this.numberOfGuests = numberOfGuests;
-    }
-
-    public int getNumberOfGuests() {
-        return numberOfGuests;
-    }
-}
